+++ src/html5-qrcode.ts
@@ -20,7 +20,7 @@ import {
     Html5QrcodeResultFactory,
     Html5QrcodeErrorFactory,
     Html5QrcodeSupportedFormats,
-    QrcodeDecoderAsync,
+    RobustQrcodeDecoderAsync,
     isValidHtml5QrcodeSupportedFormats,
     Html5QrcodeConstants,
     Html5QrcodeResult,
@@ -55,6 +55,7 @@ class Constants extends Html5QrcodeConstants {
     static DEFAULT_WIDTH = 300;
     static DEFAULT_WIDTH_OFFSET = 2;
     static FILE_SCAN_MIN_HEIGHT = 300;
+    static FILE_SCAN_HIDDEN_CANVAS_PADDING = 100;
     static MIN_QR_BOX_SIZE = 50;
     static SHADED_LEFT = 1;
     static SHADED_RIGHT = 2;
@@ -87,7 +88,7 @@ export interface Html5QrcodeConfigs {
      * enable faster native code scanning experience.
      * 
      * Set this flag to true, to enable using {@class BarcodeDetector} if
-     * supported. This is false by default.
+     * supported. This is true by default.
      * 
      * Documentations:
      *  - https://developer.mozilla.org/en-US/docs/Web/API/BarcodeDetector
@@ -255,7 +256,7 @@ export class Html5Qrcode {
     private readonly logger: Logger;
     private readonly elementId: string;
     private readonly verbose: boolean;
-    private readonly qrcode: QrcodeDecoderAsync;
+    private readonly qrcode: RobustQrcodeDecoderAsync;
 
     private shouldScan: boolean;
 
@@ -681,27 +682,41 @@ export class Html5Qrcode {
                         /* dHeight= */ config.height);
                 }
 
+                // Hidden canvas should be at-least as big as the image.
+                // This could get really troublesome for large images like 12MP
+                // images or 48MP images captured on phone.
+                let padding = Constants.FILE_SCAN_HIDDEN_CANVAS_PADDING;
+                let hiddenImageWidth = Math.max(inputImage.width, config.width);
+                let hiddenImageHeight = Math.max(inputImage.height, config.height);
+
+                let hiddenCanvasWidth = hiddenImageWidth + 2 * padding;
+                let hiddenCanvasHeight = hiddenImageHeight + 2 * padding;
+
+                // Try harder for file scan.
+                // TODO(minhazav): Fallback to mirroring, 90 degree rotation and
+                //  color inversion.
                 const hiddenCanvas = this.createCanvasElement(
-                    config.width, config.height);
+                    hiddenCanvasWidth, hiddenCanvasHeight);
                 element.appendChild(hiddenCanvas);
                 const context = hiddenCanvas.getContext("2d");
                 if (!context) {
                     throw "Unable to get 2d context from canvas";
                 }
-                context.canvas.width = config.width;
-                context.canvas.height = config.height;
+
+                context.canvas.width = hiddenCanvasWidth;
+                context.canvas.height = hiddenCanvasHeight;
                 context.drawImage(
                     inputImage,
                     /* sx= */ 0,
                     /* sy= */ 0,
                     /* sWidth= */ imageWidth,
                     /* sHeight= */ imageHeight,
-                    /* dx= */ 0,
-                    /* dy= */  0,
-                    /* dWidth= */ config.width,
-                    /* dHeight= */ config.height);
+                    /* dx= */ padding,
+                    /* dy= */  padding,
+                    /* dWidth= */ hiddenImageWidth,
+                    /* dHeight= */ hiddenImageHeight);
                 try {
-                    this.qrcode.decodeAsync(hiddenCanvas)
+                    this.qrcode.decodeRobustlyAsync(hiddenCanvas)
                         .then((result) => {
                             resolve(
                                 Html5QrcodeResultFactory.createFromQrcodeResult(
@@ -891,26 +906,27 @@ export class Html5Qrcode {
     /*eslint complexity: ["error", 10]*/
     private getUseBarCodeDetectorIfSupported(
         config: Html5QrcodeConfigs | undefined) : boolean {
+        // Default value is true.
         if (isNullOrUndefined(config)) {
-            return false;
+            return true;
         }
 
         if (!isNullOrUndefined(config!.useBarCodeDetectorIfSupported)) {
             // Default value is false.
-            return config!.useBarCodeDetectorIfSupported === true;
+            return config!.useBarCodeDetectorIfSupported !== false;
         }
 
         if (isNullOrUndefined(config!.experimentalFeatures)) {
-            return false;
+            return true;
         }
 
         let experimentalFeatures = config!.experimentalFeatures!;
         if (isNullOrUndefined(
             experimentalFeatures.useBarCodeDetectorIfSupported)) {
-            return false;
+            return true;
         }
 
-        return experimentalFeatures.useBarCodeDetectorIfSupported === true;
+        return experimentalFeatures.useBarCodeDetectorIfSupported !== false;
     }
 
     /**